refactor: move logic and types from `routes` to `logic`
