+++ src/logic/safe/store/actions/transactions/utils/transferDetails.d.ts
